Refactor RF4S service into smaller files
Refactor the RF4S service into smaller, more manageable files to improve code organization and maintainability.
