+++ pyside6_overlay/services/__init__.py
@@ -5,7 +5,13 @@ Contains background services and integrations
 """
 
 from .rf4s_service import RF4SService
+from .rf4s_connection import RF4SConnection
+from .rf4s_commands import RF4SCommands
+from .rf4s_data_processor import RF4SDataProcessor
 
 __all__ = [
-    'RF4SService'
+    'RF4SService',
+    'RF4SConnection',
+    'RF4SCommands',
+    'RF4SDataProcessor'
 ]
